Fix email templates To/From to conform to RFC 5322
This removes the duplicate mention of the email address in the To: and
From: headers of templates, because the @ character is not allowed as
part of an atom.

It would only be allowed if the email address in the name part was quoted.

Details: http://tools.ietf.org/html/rfc5322#section-3.2.3

This fixes display problems in Apple Mail.